pyproject.toml
[project]
name = "movement"
version = "1.0.0"
description = "Add your description here"
requires-python = ">=3.9.19"
dependencies = [
    "fastapi>=0.124.0",
    "rpi-gpio>=0.7.1",
    "uvicorn>=0.38.0",
]
